# doc/api.yml
openapi: 3.0.3

info:
  title: geo-jobs
  description: Launch Zone Tiling Jobs (ZTJ) and Zone Detection Jobs (ZDJ).
  version: 'latest'

servers:
  - url: 'https://localhost:8080'


paths:
  '/ping':
    get:
      operationId: ping
      tags:
        - Health
      responses:
        200:
          description: pong
          content:
            text/plain:
              schema:
                type: string
                example: pong
  '/jobs/{id}/annotationProcessing':
    post:
      tags:
        - Tiling
        - Detection
      operationId: processAnnotationJob
      description: | 
        Send jobs tiles to bpartners-annotator.
        In case of ZDJ, detected objects are sent with all images.
        In case of ZTJ, only images are sent with empty detected objects.
      parameters:
        - in: path
          name: id
          required: true
          description: ZTJ or ZDJ identifier
          schema:
            type: string
      responses:
        200:
          description: The created ZTJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationJobProcessing"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/import':
    post:
      tags:
        - Tiling
      operationId: importZTJ
      description: Creates a ZTJ from an external source
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportZoneTilingJob"
      responses:
        200:
          description: The created ZTJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneTilingJob"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs':
    post:
      tags:
        - Tiling
      operationId: tileZone
      description: Creates a ZTJ
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneTilingJob"
      responses:
        200:
          description: The created ZTJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneTilingJob"
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Tiling
      operationId: getTilingJobs
      description: List all ZTJ
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        200:
          description: All ZTJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneTilingJob"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/{id}/taskStatistics':
    get:
      tags:
        - Tiling
      parameters:
        - in: path
          name: id
          required: true
          description: ZTJ identifier
          schema:
            type: string
      operationId: getTilingTaskStatistics
      description: Get tiling task statistics for a given ZTJ
      responses:
        200:
          description: List of tiling task statistics
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/TaskStatistic"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/{id}/taskFiltering':
    post:
      tags:
        - Tiling
      parameters:
        - in: path
          name: id
          required: true
          description: ZTJ identifier
          schema:
            type: string
      operationId: filterTilingTasks
      description: | 
        Separate SUCCEEDED tasks into new tiling job from not SUCCEEDED tasks inside another tiling job, for the given ZTJ
      responses:
        200:
          description: | 
            List of new tiling jobs with its tasks statuses.
            SUCCEEDED status value means the job contains only SUCCEEDED tasks.
            NOT_SUCCEEDED status value means the job contains only not SUCCEEDED tasks, i.e FAILED or UNKNOWN
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilteredTilingJob"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/{id}/duplications':
    post:
      tags:
        - Tiling
      parameters:
        - in: path
          name: id
          required: true
          description: ZTJ identifier
          schema:
            type: string
      operationId: duplicateTilingJob
      description: Duplicate existing zone tiling job with its tiling tasks and tiles
      responses:
        200:
          description: Duplicated tiling job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneTilingJob"
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/{id}/retry':
    put:
      tags:
        - Tiling
      operationId: processFailedTilingJob
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ZTJ Identifier
          required: true
      description: Retry failed tasks for a tiling job
      responses:
        200:
          description: The ZTJ with RETRYING health and PROCESSING progression status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneTilingJob"
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/tilingJobs/{id}/parcels':
    get:
      tags:
        - Tiling
      parameters:
        - in: path
          name: id
          required: true
          description: ZTJ identifier
          schema:
            type: string
      operationId: getZTJParcels
      description: List all parcels from a ZTJ
      responses:
        200:
          description: All parcels for given ZTJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parcel"
        '500':
          $ref: '#/components/responses/500'

  '/detectionJobs':
    get:
      tags:
        - Detection
      operationId: getDetectionJobs
      description: List all ZDJ
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        200:
          description: All ZDJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/retry':
    put:
      tags:
        - Tiling
      operationId: processFailedDetectionJob
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ZDJ Identifier
          required: true
      description: Retry failed tasks for a detection job
      responses:
        200:
          description: The ZDJ with RETRYING health and PROCESSING progression status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetectionJob"
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/detectedParcels':
    get:
      tags:
        - Detection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJParcels
      description: List all detected parcels from a ZTJ
      responses:
        200:
          description: All detected parcels for given ZTJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetectedParcel"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/geojsonsUrl':
    get:
      tags:
        - Detection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJGeojsonsUrl
      description: |
        Get a secured URL to a ZIP containing the detection results as GeoJSON,
        one GeoJSON per objectType for the whole job.
      responses:
        200:
          description: Information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoJsonsUrl'
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/process':
    post:
      tags:
        - Detection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: processZDJ
      description: Process a ZDJ
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DetectableObjectConfiguration"
      responses:
        200:
          description: The processed (-ing) ZDJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'

components:
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    429:
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    CreateZoneTilingJob:
      type: object
      properties:
        emailReceiver:
          type: string
        zoneName:
          type: string
        geoServerUrl:
          type: string
        geoServerParameter:
          $ref: "#/components/schemas/GeoServerParameter"
        features:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
        zoomLevel:
          type: string
          enum:
            - WORLD_0
            - WORLD_1
            - WORLD_2
            - CONTINENT_0
            - CONTINENT_1
            - COUNTRIES
            - COUNTRY
            - STATES
            - COUNTIES_0
            - COUNTIES_1
            - COUNTY
            - METROPOLITAN_AREA
            - CITIES
            - CITY
            - TOWN
            - NEIGHBORHOOD
            - STREETS
            - CITY_BLOCK
            - BUILDINGS
            - BUILDING
            - HOUSES_0
            - HOUSES_1
            - HOUSES_2
            - HOUSE_PROPERTY
    AnnotationJobProcessing:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        annotationWithObjectTruePositive:
          type: string
        annotationWithObjectFalsePositive:
          type: string
        annotationWithoutObjectJobId:
          type: string
        creationDatetime:
          type: string
          format: date-time
    FilteredTilingJob:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/SuccessStatus'
        job:
          $ref: '#/components/schemas/ZoneTilingJob'
    ImportZoneTilingJob:
      type: object
      properties:
        createZoneTilingJob:
          $ref: '#/components/schemas/CreateZoneTilingJob'
        #TODO: must be oneOf but for now, only from S3 bucket is implemented
        bucketName:
          type: string
        bucketPathPrefix:
          type: string
    ZoneTilingJob:
      allOf:
        - $ref: "#/components/schemas/CreateZoneTilingJob"
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/Status"

    CreateZoneDetectionJob:
      type: object
      properties:
        emailReceiver:
          type: string
        zoneName:
          type: string
        zoneTilingJobId:
          type: string
        type:
          $ref: '#/components/schemas/ZoneDetectionType'
        objectsToDetect:
          type: array
          items:
           $ref: "#/components/schemas/DetectableObjectConfiguration"
    ZoneDetectionJob:
      allOf:
        - $ref: "#/components/schemas/CreateZoneDetectionJob"
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/Status"
    ZoneDetectionType:
      enum:
        - MACHINE
        - HUMAN
    DetectableObjectConfiguration:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/DetectableObjectType"
        confidence:
          type: number
          description: between 0 and 1
    DetectableObjectType:
      type: string
      enum:
        - ROOF
        - SOLAR_PANEL
        - POOL
        - PATHWAY
        - TREE
        - SIDEWALK
        - LINE
        - GREEN_SPACE

    Parcel:
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        feature:
          $ref: "#/components/schemas/Feature"
        tiles:
          type: array
          items:
            $ref: "#/components/schemas/Tile"
        tilingStatus:
          $ref: "#/components/schemas/Status"
    DetectedParcel:
      properties:
        id:
          type: string
        detectionJobIb:
          type: string
        parcelId:
          type: string
        creationDatetime:
          type: string
          format: date-time
        detectedTiles:
          type: array
          items:
            $ref: "#/components/schemas/DetectedTile"
        status:
          $ref: "#/components/schemas/Status"
    TaskStatistic:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        actualJobStatus:
          $ref: '#/components/schemas/Status'
        taskStatusStatistics:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusStatistic'
        updatedAt:
          type: string
          format: date-time
    TaskStatusStatistic:
      type: object
      properties:
        progression:
          $ref: '#/components/schemas/ProgressionStatus'
        healthStatistics:
          type: array
          items:
            $ref: '#/components/schemas/HealthStatusStatistic'
    HealthStatusStatistic:
      type: object
      properties:
        health:
          $ref: '#/components/schemas/HealthStatus'
        count:
          type: number
    SuccessStatus:
      type: string
      enum:
        - SUCCEEDED
        - NOT_SUCCEEDED
    Tile:
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        coordinates:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
            z:
              type: integer
        bucketPath:
          type: string
    DetectedTile:
      properties:
        tileId:
          type: string
        creationDatetime:
          type: string
          format: date-time
        detectedObjects:
          type: array
          items:
              $ref: "#/components/schemas/DetectedObject"
        status:
          $ref: "#/components/schemas/Status"
        bucketPath:
          type: string
    DetectedObject:
      properties:
          detectedObjectType:
            $ref: "#/components/schemas/DetectableObjectType"
          feature:
            $ref: "#/components/schemas/Feature"
          confidence:
            type: number
          detectorVersion:
            type: string
    JobType:
      type: string
      enum:
        - DETECTION
        - TILING
    ProgressionStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - FINISHED
    HealthStatus:
      type: string
      enum:
        - SUCCEEDED
        - FAILED
        - UNKNOWN
        - RETRYING
    Status:
      type: object
      properties:
        progression:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - FINISHED
        health:
          type: string
          enum:
            - SUCCEEDED
            - FAILED
            - UNKNOWN
            - RETRYING
        creationDatetime:
          type: string
          format: date-time

    ###########################################################################################################
    ###  GeoJSON objects. Modified from: https://gist.github.com/bubbobne/fe5f2db65acf039be6a9fd92fc9c7233  ###
    ###########################################################################################################

    GeoServerParameter:
      type: object
      properties:
        service:
          type: string
        request:
          type: string
        layers:
          type: string
        styles:
          type: string
        format:
          type: string
        transparent:
          type: boolean
        version:
          type: string
        width:
          type: integer
        height:
          type: integer
        srs:
          type: string

    Feature:
      type: object
      properties:
        id:
          type: string
        zoom:
          type: integer
        geometry:
          $ref: "#/components/schemas/MultiPolygon"

    Geometry:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - Point
            - LineString
            - Polygon
            - MultiPolygon

    Point2D:
      type: array
      maxItems: 2
      minItems: 2
      items:
        type: number

    Point:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          $ref: '#/components/schemas/Point2D'

    Polygon:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Point2D'

    MultiPolygon:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                $ref: '#/components/schemas/Point2D'
    GeoJsonsUrl:
      type: object
      properties:
        url:
          type: string
        status:
          $ref: "#/components/schemas/Status"